change protocol version type to u32

+++ packages/rs-dpp/src/errors/compatible_protocol_version_is_not_defined_error.rs
@@ -3,17 +3,17 @@ use thiserror::Error;
 #[derive(Error, Debug)]
 #[error("Compatible version is not defined for protocol version {current_protocol_version}")]
 pub struct CompatibleProtocolVersionIsNotDefinedError {
-    current_protocol_version: u64,
+    current_protocol_version: u32,
 }
 
 impl CompatibleProtocolVersionIsNotDefinedError {
-    pub fn new(current_protocol_version: u64) -> Self {
+    pub fn new(current_protocol_version: u32) -> Self {
         Self {
             current_protocol_version,
         }
     }
 
-    pub fn current_protocol_version(&self) -> u64 {
+    pub fn current_protocol_version(&self) -> u32 {
         self.current_protocol_version
     }
 }
